% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability_function.R
\name{Reliability_test}
\alias{Reliability_test}
\title{Reliability_test() Function}
\usage{
Reliability_test(x, t)
}
\description{
This function mean
}
\examples{
Reliability_test()
}
\keyword{Reliability_test}
